requirements.txt
geoip2>=4.8.0
user-agents>=2.2.0
pyyaml>=6.0
